requirements.txt
jaeger-client==3.12.0
python-json-logger==0.1.10
PyYAML==3.13
anyconfig==0.9.7
requests==2.20.0
# Optionals:
connexion[swagger-ui]==2.0.2
# Necessary to deploy swagger ui files in projects that install this library,
# connexion's extra it isn't working as expected in that case
swagger-ui-bundle==0.0.2
flask==1.0.2
Flask-OpenTracing==0.2.0

